+++ gridtools/topoutils.py
@@ -1,0 +1,4 @@
+# Topography tools
+# Implementations of topographic grid generators
+#  - https://github.com/nikizadehgfdl/ocean_model_topog_generator.git
+#    - generate_h2_by_objective_regridding_of_high_resolution_topography(**opts)